Display weather data based on searched place
The user is able to search a place by writing
into the input field and then search by pressing
the search button.

The place shown depends on the searching algorithm
of Visual Crossing. Also when no valid place is
returned then a error message is displayed to the
user (either invalid or empty)